Move `maskEnvVar()` to `helpers/utils.ts`
